+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 - Allow for broader range of number of attempts, including unlimited, for scored pages
 - Add survey authoring support (behind feature flag)
+- Add File Upload activity
 
 ## 0.19.3 (2022-05-17)